+++ packages/shared-src/src/reports/index.js
@@ -20,6 +20,7 @@ import * as covidSwabLabTestsSummary from './covid-module/covid-swab-lab-tests-s
 import * as indiaAssistiveTechnologyDeviceLineList from './india-assistive-technology-device-line-list';
 import * as iraqAssistiveTechnologyDeviceLineList from './iraq-assistive-technology-device-line-list';
 import * as pngAssistiveTechnologyDeviceLineList from './png-assistive-technology-device-line-list';
+import * as fijiAspenEncounterSummaryLineList from './fiji-aspen-encounter-summary-line-list';
 import * as fijiRecentAttendanceList from './fiji-recent-attendance-list';
 import * as fijiNcdPrimaryScreeningLineList from './fiji-ncd-primary-screening/fiji-ncd-primary-screening-line-list';
 import * as fijiNcdPrimaryScreeningPendingReferralsLineList from './fiji-ncd-primary-screening/fiji-ncd-primary-screening-pending-referrals-line-list';
@@ -73,6 +74,8 @@ export function getReportModule(reportType) {
       return iraqAssistiveTechnologyDeviceLineList;
     case 'png-assistive-technology-device-line-list':
       return pngAssistiveTechnologyDeviceLineList;
+    case 'fiji-aspen-encounter-summary-line-list':
+      return fijiAspenEncounterSummaryLineList;
     case 'fiji-recent-attendance-list':
       return fijiRecentAttendanceList;
     case 'fiji-ncd-primary-screening-line-list':